requirements.txt
Flask==2.3.3
openai>=1.78.0
PyPDF2>=3.0.1
python-dotenv>=1.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.1.1
Werkzeug>=2.3.7
psycopg2-binary>=2.9.7
Flask-Limiter>=3.5.0 